refactor(cypress-integration): Renamed /nextjs_auth0/ folder to /auth-js/
