+++ integrations/cypress/cypress/support/commands/auth0-js/auth.ts
